Adds duration setting slider

+++ app/src/main/java/sporttimer/minibilles/fr/sporttimer/Timer.kt
@@ -2,7 +2,7 @@ package sporttimer.minibilles.fr.sporttimer
 
 data class TimerDescription(
         var name: String = "New timer",
-        var durations: MutableList<Double> = arrayListOf(30.0, 30.0),
+        var durations: MutableList<Int> = arrayListOf(30, 30),
 
         var currentTime: Double = 0.0,
         var globalCount: Int = 0,
